Struct copper::lsm_tree::sstable::SSTable

pub struct SSTable { /* private fields */ }

Implementations

impl SSTable

pub fn get(&self, key: &[u8]) -> Option<Vec<u8>>

Returns the value of the entry with the given key if it exists and the bloom filter indicates that it might be in the SSTable.

Arguments

key - A byte slice that holds the key of the entry.

pub fn from_memtable(memtable: &Memtable) -> Self

Creates a new SSTable from a Memtable.

Arguments

memtable - A reference to a Memtable that holds the entries to be included in the SSTable.

The function creates a new bloom filter and a new BTreeMap of entries, and inserts all entries from the Memtable into the BTreeMap and the bloom filter.

pub fn save_to_disk(&self, path: &str) -> Result<(), Error>

Saves the SSTable to a file on disk.

Arguments

path - A string slice that specifies the path of the file.

The function creates a new file at the specified path, and writes the number of entries in the SSTable and each key-value pair to the file.

pub fn get_all_entries(&self) -> Vec<Entry>

Saves the SSTable to a file on disk.

Arguments

path - A string slice that specifies the path of the file.

The function creates a new file at the specified path, and writes the number of entries in the SSTable and each key-value pair to the file.

pub fn load_from_disk(path: &str) -> Result<Self, Error>

Loads an SSTable from a file on disk.

Arguments

path - A string slice that specifies the path of the file.

The function opens the file at the specified path, reads the number of entries in the SSTable and each key-value pair from the file, and creates a new SSTable with the read entries and a new bloom filter.

pub fn get_size(&self) -> usize

Returns the size of the SSTable, which is the sum of the lengths of all keys, values, and tombstones.

pub fn compact(level: &Vec<Self>) -> Self

Compacts a level of SSTables into a single SSTable.

Arguments

level - A reference to a vector of SSTables that represents the level to be compacted.

The function creates a new SSTable and a new bloom filter, and inserts all entries from the SSTables in the level into the SSTable and the bloom filter.

pub fn get_range(&self, start: &[u8], end: &[u8]) -> Vec<Entry>

Returns a range of entries in the SSTable from the start key to the end key, inclusive.

Arguments

start - A byte slice that holds the start key of the range.

end - A byte slice that holds the end key of the range.

The function does not return deleted entries.

pub fn clear(&mut self)

Clears the SSTable, removing all entries and clearing the bloom filter.

pub fn merge(sstables: &[SSTable]) -> SSTable

Merges multiple SSTables into a single SSTable.

Arguments

sstables - A slice of SSTables that are to be merged.

The function creates a new SSTable and a new bloom filter, and inserts all entries from the SSTables into the SSTable and the bloom filter.